Moved NicetyDisplay into its own file
