+++ src/components/NicetyDisplay.js
@@ -1,0 +1,43 @@
+import React from 'react';
+import { Grid } from 'react-bootstrap';
+
+import NicetyRow from './NicetyRow';
+
+const NicetyDisplay = React.createClass({
+
+    getInitialState: function() {
+        return {
+            niceties: []
+        };
+    },
+
+    generateRows: function() {
+        let dataList = [];
+        for (let i = 0; i < this.props.niceties.length; i +=4) {
+            let row = [];
+            for (let j = 0; j < 4; j++) {
+                if ((i + j) < this.props.niceties.length) {
+                    row.push(this.props.niceties[i + j]);
+                }
+            }
+            dataList.push(row);
+        }
+        return dataList;
+    },
+
+    render: function() {
+        let list = this.generateRows();
+        return (
+            <div className="niceties">
+              <Grid>
+                {list.map(function(row) {
+                    return (
+                        <NicetyRow data={row}/>
+                    );
+                })}
+            </Grid>
+                </div>
+        );}
+});
+
+export default NicetyDisplay;
